# contracts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "BuildSnapshotRequest": {
        "description": "Запит на побудову зрізу.",
        "properties": {
          "periodKey": {
            "description": "Період.",
            "format": "int32",
            "type": "integer"
          },
          "projectId": {
            "description": "Проєкт.",
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "projectId",
          "periodKey"
        ],
        "type": "object"
      },
...
      "CloneProjectRequest": {
        "description": "Запит на клонування проєкту.",
        "properties": {
          "code": {
            "description": "Код нового проєкту.",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "type": "object"
      },
...
      "CreateProjectRequest": {
        "description": "Запит на створення проєкту.",
        "properties": {
          "code": {
            "description": "Код проєкту.",
            "type": "string"
          },
          "nameL10n": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Назва мовами каталогу.",
            "type": "object"
          },
          "periodKind": {
            "description": "Періодичність.",
            "type": "string"
          },
          "periodPolicyId": {
            "description": "Політика зсувів періодів.",
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          },
          "templateVersionId": {
            "description": "Версія шаблону, за якою заповнюються документи.",
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          },
          "timeZoneId": {
            "description": "Пояс майданчика; після відкриття періоду не змінюється.",
            "type": "string"
          },
          "year": {
            "description": "Звітний рік; типово поточний.",
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          }
        },
        "required": [
          "code",
          "nameL10n",
          "timeZoneId",
          "periodKind"
        ],
        "type": "object"
      },
...
      "CurrentPeriodMode": {
        "description": "Режим визначення поточного періоду проєкту (D-77).",
        "enum": [
          "Auto",
          "Pinned"
        ]
      },
...
      "JobAcceptedResponse": {
        "description": "Тіла відповідей, спільні для кількох контролерів.",
        "properties": {
          "jobId": {
            "description": "Ідентифікатор фонової задачі для опитування стану.",
            "type": "string"
          }
        },
        "required": [
          "jobId"
        ],
        "type": "object"
      },
...
      "PagedResultOfProjectSummary": {
        "description": "Сторінка результатів. Ендпоінтів, що повертають «усе», не існує —\r\nперевіряється архітектурним тестом.",
        "properties": {
          "items": {
            "description": "Елементи сторінки.",
            "items": {
              "$ref": "#/components/schemas/ProjectSummary"
            },
            "type": "array"
          },
          "nextCursor": {
            "description": "Курсор наступної сторінки; `null` — кінець.",
            "type": [
              "null",
              "string"
            ]
          },
          "totalCount": {
            "description": "Загальна кількість; `null`, якщо підрахунок дорогий.",
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor",
          "totalCount"
        ],
        "type": "object"
      },
...
      "PeriodCalendarDto": {
        "description": "Календар проєкту.",
        "properties": {
          "currentPeriodMode": {
            "$ref": "#/components/schemas/CurrentPeriodMode",
            "description": "Автоматичний вибір чи закріплений період."
          },
          "periodKind": {
            "$ref": "#/components/schemas/PeriodKind",
            "description": "Періодичність."
          },
          "periods": {
            "description": "Періоди в порядку зростання `PeriodKey`.",
            "items": {
              "$ref": "#/components/schemas/PeriodDto"
            },
            "type": "array"
          },
          "projectId": {
            "description": "Проєкт.",
            "format": "int32",
            "type": "integer"
          },
          "timeZoneId": {
            "description": "Пояс майданчика, у якому пораховані межі.",
            "type": "string"
          }
        },
        "required": [
          "projectId",
          "timeZoneId",
          "periodKind",
          "currentPeriodMode",
          "periods"
        ],
        "type": "object"
      },
      "PeriodDto": {
        "description": "Період у календарі проєкту.",
        "properties": {
          "endsAt": {
            "description": "Кінець періоду в поясі майданчика, виключно.",
            "format": "date-time",
            "type": "string"
          },
          "graceEndsAt": {
            "description": "Кінець пільгового вікна; після нього період закривається.",
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "id": {
            "description": "Ідентифікатор періоду.",
            "format": "int32",
            "type": "integer"
          },
          "isCurrent": {
            "description": "Чи є періодом за замовчуванням для UI.",
            "type": "boolean"
          },
          "periodKey": {
            "description": "`Year*100 + Sequence`; він же ключ партиції.",
            "format": "int32",
            "type": "integer"
          },
          "reopenedUntil": {
            "description": "Якщо період відкрито повторно — до якого моменту.",
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "sequence": {
            "description": "Порядковий номер у році: `1…12` для місячних, `1…4` для квартальних.",
            "format": "int32",
            "type": "integer"
          },
          "startsAt": {
            "description": "Початок періоду в поясі майданчика.",
            "format": "date-time",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/PeriodState",
            "description": "Стан; обчислює `PeriodStateJob`, а не запит."
          },
          "year": {
            "description": "Рік.",
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "periodKey",
          "year",
          "sequence",
          "state",
          "startsAt",
          "endsAt",
          "graceEndsAt",
          "isCurrent",
          "reopenedUntil"
        ],
        "type": "object"
      },
      "PeriodKind": {
        "description": "Гранулярність періоду проєкту.",
        "enum": [
          "Monthly",
          "Quarterly",
          "Yearly",
          "Custom"
        ]
      },
      "PeriodState": {
        "description": "Стан звітного періоду. Обчислює `PeriodStateJob`, а не запит.",
        "enum": [
          "Scheduled",
          "Open",
          "Grace",
          "Closed"
        ]
      },
...
      "ProblemDetails": {
        "properties": {
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
      "ProjectIdResponse": {
        "description": "Створений проєкт.",
        "properties": {
          "projectId": {
            "description": "Ідентифікатор.",
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "projectId"
        ],
        "type": "object"
      },
      "ProjectStatus": {
        "description": "Стан проєкту: `Draft → Active → Archived`.",
        "enum": [
          "Draft",
          "Active",
          "Archived"
        ]
      },
      "ProjectSummary": {
        "description": "Проєкт у переліку.",
        "properties": {
          "code": {
            "description": "Код.",
            "type": "string"
          },
          "currentPeriodId": {
            "description": "Поточний період — підказка UI, не правило доступу (D-77).",
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          },
          "id": {
            "description": "Ідентифікатор.",
            "format": "int32",
            "type": "integer"
          },
          "periodCount": {
            "description": "Скільки періодів у календарі.",
            "format": "int32",
            "type": "integer"
          },
          "periodKind": {
            "$ref": "#/components/schemas/PeriodKind",
            "description": "Періодичність."
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus",
            "description": "Стан проєкту."
          },
          "timeZoneId": {
            "description": "Пояс майданчика.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "status",
          "timeZoneId",
          "periodKind",
          "currentPeriodId",
          "periodCount"
        ],
        "type": "object"
      },
...
      "ReportSnapshotSummary": {
        "description": "Зріз у переліку.",
        "properties": {
          "builtAt": {
            "description": "Коли побудовано.",
            "format": "date-time",
            "type": "string"
          },
          "contentHash": {
            "description": "Контрольна сума вмісту в hex; `null` — не рахувалася.",
            "type": [
              "null",
              "string"
            ]
          },
          "id": {
            "description": "Ідентифікатор зрізу.",
            "format": "int64",
            "type": "integer"
          },
          "isCurrent": {
            "description": "Чи це поточний зріз для пари «версія × проєкт × період».",
            "type": "boolean"
          },
          "periodKey": {
            "description": "Період; `null` — увесь рік проєкту.",
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          },
          "projectId": {
            "description": "Проєкт.",
            "format": "int32",
            "type": "integer"
          },
          "reportVersionId": {
            "description": "Версія звіту.",
            "format": "int32",
            "type": "integer"
          },
          "rowCount": {
            "description": "Скільки рядків.",
            "format": "int32",
            "type": "integer"
          },
          "status": {
            "description": "Статус даних зрізу (D-65).",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reportVersionId",
          "projectId",
          "periodKey",
          "status",
          "isCurrent",
          "rowCount",
          "contentHash",
          "builtAt"
        ],
        "type": "object"
      },
...
      "SetCurrentPeriodRequest": {
        "description": "Запит на зміну поточного періоду.",
        "properties": {
          "pinnedPeriodId": {
            "description": "Закріплений період; `null` — режим `Auto`.",
            "format": "int32",
            "type": [
              "null",
              "integer"
            ]
          },
          "reason": {
            "description": "Причина закріплення; потрапляє в аудит.",
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "pinnedPeriodId",
          "reason"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Ecr.Api | v1"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfProjectSummary"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfProjectSummary"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfProjectSummary"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Перелік проєктів. Право `Document.View`.",
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "description": "`TimeZoneId` задається тут і **не змінюється** після відкриття\r\nпершого періоду (`ECR-PRD-0409`): межі періодів рахуються в поясі\r\nмайданчика, і зміна поясу заднім числом зсунула б уже подану звітність.",
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectIdResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectIdResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectIdResponse"
                }
              }
            },
            "description": "Created"
          }
        },
        "summary": "Створює проєкт. Право `Project.Manage`.",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/v1/projects/{id}/activate": {
      "post": {
        "description": "⛔ Без цього маршруту проєкт лишається чернеткою назавжди, періоди не\r\nвідкриваються і система не приймає жодного значення (`A7-25`).",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Активує проєкт. Право `Project.Manage`.",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/v1/projects/{id}/archive": {
      "post": {
        "description": "⚠ Це позначка, а не перенесення даних: фізично в `arc.*` їх\r\nпереносить окрема задача, і робить це свідомим кроком людини.\r\nДозволено лише коли всі періоди закриті (`D-123`).",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Позначає проєкт заархівованим. Право `Project.Manage`.",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/v1/projects/{id}/clone": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloneProjectRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloneProjectRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloneProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectIdResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectIdResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectIdResponse"
                }
              }
            },
            "description": "Created"
          }
        },
        "summary": "Клонує проєкт разом із налаштуваннями. Право `Project.Manage`.",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/v1/projects/{id}/current-period": {
      "put": {
        "description": "⚠ `CurrentPeriod` — підказка UI, а **не** правило доступу\r\n(`D-77`): на рішення про право запису він не впливає взагалі.\r\n`null` у `PinnedPeriodId` повертає режим `Auto`.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetCurrentPeriodRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCurrentPeriodRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCurrentPeriodRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Встановлює поточний період проєкту. Право `Period.Configure`.",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/v1/projects/{id}/periods": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeriodCalendarDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeriodCalendarDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PeriodCalendarDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Календар періодів проєкту. Право `Document.View`.",
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/api/v1/reports/snapshots": {
      "get": {
        "description": "У переліку є час побудови й контрольна сума: споживач має бачити, на\r\nяких даних побудовано зріз. Два зрізи однієї версії за один період\r\nвідрізняються лише цим.",
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "periodKey",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReportSnapshotSummary"
                  },
                  "type": "array"
                }
              },
              "text/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReportSnapshotSummary"
                  },
                  "type": "array"
                }
              },
              "text/plain": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReportSnapshotSummary"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Перелік побудованих зрізів. Право `Report.ViewRegulatory`.",
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/v1/reports/{code}/build": {
      "post": {
        "description": "Зріз **незмінний**: повторна побудова створює новий, а не переписує\r\nстарий. Інакше звіт, роздрукований учора, і той самий звіт сьогодні\r\nдавали б різні числа без жодного сліду.",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BuildSnapshotRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildSnapshotRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildSnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Accepted"
          }
        },
        "summary": "Будує зріз. Право `Report.BuildSnapshot`.",
        "tags": [
          "Reports"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
...
}